+++ CHANGELOG.md
@@ -3,11 +3,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
+
+## [3.0.1] - 2020-08-27
+
+### Fixed
+- Improved speed of query for `GET /api/trends/region/:attribute`.
+  [GEOD-1343](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1343)
  
 ## [3.0.0] - 2020-02-27
 *Warning*: Requires Postgres schema update. See `geostreams.sql`.
-### Added
-- Postgres view to speed up downloading of datapoints.
+
+### Fixed
+- Improved datapoints downloading speed by adding a Postgres view.
  
 ## [3.0.0-beta.5] - 2019-05-16
 ### Added
@@ -82,5 +89,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   Query to get sensor is expensive since it calculates min / max time dynamically.
 - Rename controller methods for consistency
   [GEOD-958](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-958)
-
+- Remove four controllers and routes that were not yet implemented
+  [GEOD-1309](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1309)
+  
 [3.0.0-alpha.1]: https://opensource.ncsa.illinois.edu/bitbucket/projects/GEOD/repos/geo-temporal-api-v2/browse